requirements.txt
Flask>=2.0.2
gunicorn>=20.1.0
Flask-hCaptcha>=0.6.0 
python-dotenv>=0.20.0
